feat: make responsive, add admin panel, rename to PandaPost, and update legal pages for TikTok compliance

+++ config/cashier.php
@@ -135,6 +135,6 @@ return [
     |
     */
 
-    'trial_days' => env('CASHIER_TRIAL_DAYS', 7),
+    'trial_days' => (int) env('CASHIER_TRIAL_DAYS', 7),
 
 ];
